LayoutTests/svg/dynamic-updates/SVGEllipseElement-svgdom-ry-prop.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="resources/SVGTestCase.js"></script>
<script src="../../resources/js-test-pre.js"></script>
<script src="../../fast/repaint/resources/repaint.js"></script>
</head>
<body onload="runRepaintTest()">
<h1>SVG 1.1 dynamic update tests</h1>
<p id="description"></p>
<div id="console"></div>
<script>
// [Name] SVGEllipseElement-svgdom-ry-prop.js
// [Expected rendering result] green ellipse with rx = ry (in fact a circle)- and a series of PASS messages

description("Tests dynamic updates of the 'ry' property of the SVGEllipseElement object")
createSVGTestCase();

var ellipseElement = createSVGElement("ellipse");
ellipseElement.setAttribute("cx", "150");
ellipseElement.setAttribute("cy", "150");
ellipseElement.setAttribute("rx", "150");
ellipseElement.setAttribute("ry", "10");
ellipseElement.setAttribute("fill", "green");

rootSVGElement.appendChild(ellipseElement);
shouldBe("ellipseElement.ry.baseVal.value", "10");

function repaintTest() {
    ellipseElement.ry.baseVal.value = 150;
    shouldBe("ellipseElement.ry.baseVal.value", "150");

    completeTest();
}

var successfullyParsed = true;
</script>
</body>
</html>
